refactor: variable and file renaming
