+++ packages/server/controllers/ConversationController.ts
@@ -1,11 +1,11 @@
 import { Request, Response, NextFunction } from 'express';
 
-import { HttpError } from '../models/error';
-import { saveConversation } from '../service/conversation/saveConversation';
-import { getUserConversations } from '../service/conversation/getUserConversations';
-import { deleteUserConversation } from '../service/conversation/deleteUserConversation';
+import { HttpError } from '../models/ErrorModel';
+import { saveConversation } from '../service/conversation/SaveConversation';
+import { getAllUserConversationsById } from '../service/conversation/GetAllUserConversationsById';
+import { deleteConversation } from '../service/conversation/DeleteConversation';
 
-export const createConvo = async (
+export const onSaveConversation = async (
   req: Request,
   res: Response,
   next: NextFunction
@@ -26,14 +26,14 @@ export const createConvo = async (
   }
 };
 
-export const getConvoOfUser = async (
+export const onGetAllUserConversationById = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
   const userId = req.params.userId;
   try {
-    const userConversations = await getUserConversations(userId);
+    const userConversations = await getAllUserConversationsById(userId);
 
     res.json({ userConversations });
   } catch (err) {
@@ -43,14 +43,14 @@ export const getConvoOfUser = async (
   }
 };
 
-export const deleteConversation = async (
+export const onDeleteConversation = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
   const conversationId = req.params.convoId;
   try {
-    await deleteUserConversation(conversationId);
+    await deleteConversation(conversationId);
 
     res.json({ message: 'Successfully deleted conversation' });
   } catch (err) {
