+++ packages/server/controllers/DirectMessageController.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 
-import { HttpError } from '../models/error';
+import { HttpError } from '../models/ErrorModel';
 import { io } from '../socket';
 import { saveMessage } from '../service/message/saveMessage';
 import { getAllConversationMessages } from '../service/message/getAllConversationMessages';
